% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_NAV.R
\name{calc_NAV}
\alias{calc_NAV}
\title{Calculate Fund NAV from Ticker NAV Info}
\usage{
calc_NAV(Ticker.NAV)
}
\arguments{
\item{Ticker.NAV}{dataframe with NAV and Date informaion}
}
\description{
Calculate Fund NAV from Ticker NAV Info
}
