% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_return_series.R
\name{calc_return_series}
\alias{calc_return_series}
\title{Use stock holdings and price data to find daily return on fund using
only invested asset. Result is produced for indivisual tickers.}
\usage{
calc_return_series(return_level = c("ticker", "fund"), benchmark = "_SPXT")
}
\arguments{
\item{return_level}{* ticker: return info for each ticker in a long format data.frame
* fund: fund return info}
}
\description{
Use stock holdings and price data to find daily return on fund using
only invested asset. Result is produced for indivisual tickers.
}
