% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary.R
\name{calc_summary}
\alias{calc_summary}
\title{calculate performance summary}
\usage{
calc_summary(returns.data, summary_type = c("quick", "CAPM", "Drawdown"))
}
\arguments{
\item{summary_type}{type of summary}

\item{return.data}{dataframe of return series with Date, Return and
Strategy type.}
}
\description{
calculate performance summary
}
