% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_total_return.R
\name{calc_total_return}
\alias{calc_total_return}
\title{calculate the cumulative return from daily or monthly returns}
\usage{
calc_total_return(data.return, return_level = c("ticker", "fund"))
}
\arguments{
\item{data.return}{a daily or monthly return series}
}
\description{
calculate the cumulative return from daily or monthly returns
}
