% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_return_series.R
\name{create_return_series}
\alias{create_return_series}
\title{Create a daily return series.}
\usage{
create_return_series()
}
\description{
Create a daily return series based on the return on the historical date
instead of the current NAV.
}
