% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_xts.R
\name{df_to_xts}
\alias{df_to_xts}
\title{Convert Dataframe with `Date`` column to an xts}
\usage{
df_to_xts(dataframe)
}
\arguments{
\item{dataframe}{data.frame object}
}
\description{
Convert Dataframe with `Date`` column to an xts
}
