% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_yahoo.R
\name{format_yahoo}
\alias{format_yahoo}
\title{Process yahoo finance csv to conform to other price information's format.}
\usage{
format_yahoo(ticker)
}
\arguments{
\item{yahoo_finance_file}{file_name of the csv file}
}
\description{
Process yahoo finance csv to conform to other price information's format.
}
