% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_trade.R
\name{log_trade}
\alias{log_trade}
\title{Log the trade to trade log}
\usage{
log_trade(ticker, price, trade_quantity, trade_type, trade_date = Sys.Date())
}
\arguments{
\item{ticker}{Ticker of the stock.}

\item{price}{The price at which the trade is executed at.}

\item{trade_quantity}{The quantity traded.}

\item{trade_type}{Buy/Sell.}

\item{trade_date}{The date that the trade is executed on.}
}
\description{
Trade log is used for all other analysis. Examples are like creating a
track record series, analyzing returns etc.
}
