% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asset_return.R
\name{plot_asset_return}
\alias{plot_asset_return}
\title{Plot each individual's total return since inception}
\usage{
plot_asset_return(ticker.list = NULL)
}
\arguments{
\item{ticker.list}{tickers to plot, if not ALL}
}
\description{
Plot each individual's total return since inception
}
