% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vami.R
\name{plot_vami}
\alias{plot_vami}
\title{Plot cumulative return of the fund}
\usage{
plot_vami(data.return, benchmark = TRUE)
}
\arguments{
\item{data.return}{Daily or monthly return series}

\item{benchmark}{If TRUE, plot the benchmark return during the same period
on the same graph.}
}
\description{
Plot cumulative return of the fund
}
