% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pkg_options.R
\name{set_pkg_options}
\alias{set_pkg_options}
\title{Set or get options for my package}
\usage{
set_pkg_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{wd}}{(\code{string};1) The working directory of all data. }
}
}

