% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_benchmark.R
\name{update_benchmark}
\alias{update_benchmark}
\title{Get benchmark prcie data}
\usage{
update_benchmark(ticker = "^SP500TR", save_name = "SPTR")
}
\arguments{
\item{ticker}{benchmark ticker. Only tested for SPTR}
}
\description{
Get benchmark prcie data
}
