% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_price_info.R
\name{update_price_info}
\alias{update_price_info}
\title{Create a daily return series.}
\usage{
update_price_info()
}
\description{
Create a daily return series based on the return on the historical date
instead of the current NAV.
}
