% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_price_info.R
\name{update_price_info}
\alias{update_price_info}
\title{Create a daily return series.}
\usage{
update_price_info(dir_log_file = NULL)
}
\arguments{
\item{dir_log_file}{directory of the log file. If not provided, look for
log file at default location.}
}
\description{
Create a daily return series based on the return on the historical date
instead of the current NAV.
}
